requirement.txt
appdirs==1.4.3
astroid==2.3.0
attrs==19.1.0
autopep8==1.4.4
cached-property==1.5.1
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
defusedxml==0.6.0
gmail==0.6.3
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.0
googleads==21.0.0
httplib2==0.13.1
idna==2.8
isodate==0.6.0
isort==4.3.21
lazy-object-proxy==1.4.2
lxml==4.4.1
mccabe==0.6.1
oauthlib==3.1.0
pyasn1==0.4.7
pyasn1-modules==0.2.6
pycodestyle==2.5.0
pylint==2.4.1
python-dotenv==0.10.3
pytz==2019.2
PyYAML==5.1.2
requests==2.22.0
requests-oauthlib==1.2.0
requests-toolbelt==0.9.1
rsa==4.0
six==1.12.0
typed-ast==1.4.0
uritemplate==3.0.0
urllib3==1.25.6
wrapt==1.11.2
xmltodict==0.12.0
zeep==3.4.0
